requirements.txt
python-telegram-bot[job-queue]
pandas
openpyxl
python-dotenv
apscheduler